기본 게임 로직 클래스 작성 중 / etc.
1. CStatScript
2. PlayerStat
3. Inven / EquipmentScript
4. GameObject Active 관련 오버헤드 처리
etc.

+++ DX11/Source/Scripts/CharacterScript/EnemyScript/CEnemyScript.cpp
@@ -6,7 +6,7 @@
 #include "Source/ScriptMgr.h"
 
 CEnemyScript::CEnemyScript()
-    :CCharacterScript(static_cast<int>(SCRIPT_TYPE::ENEMYSCRIPT))
+    : CCharacterScript(SCRIPT_TYPE::ENEMYSCRIPT)
 {
 }
